backend/requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
aiofiles
pydantic

python-multipart
zipfile36
pyjwt
jwt-python